Build BIRD in the Dockerfile
We used to copy the BIRD executables from the local filesystem into
the docker image but that doesn't work with CI so now we get BIRD from
upstream and build it in the Docker container.

We use a git checkout instead of downloading the tarballs because the
BIRD build process assumes that it's running in a git repo and gripes
if it isn't.  This might be harmless but I'd rather have the fewest
number of warnings that I can.